rename to event_util.h
