+++ src/storage.cc
@@ -44,7 +44,7 @@
 #include "event_listener.h"
 #include "compact_filter.h"
 #include "table_properties_collector.h"
-#include "event_utils.h"
+#include "event_util.h"
 
 namespace Engine {
