+++ src/util.cc
@@ -43,7 +43,7 @@
 
 #include "util.h"
 #include "status.h"
-#include "event_utils.h"
+#include "event_util.h"
 
 #ifndef POLLIN
 # define POLLIN      0x0001    /* There is data to read */
